temp

+++ src/redux/actions/logout.js
@@ -1,14 +1,17 @@
 import actionTypes from "./actionTypes";
-
+import Reactotron from "reactotron-react-js";
 export function logout(token, socket) {
-  return async function (dispatch) {
-    await socket.emit("logout", token, async (error) => {
-      if (error) {
-        return dispatch({ type: actionTypes.LOG_OUT, logout: false });
+  return function (dispatch) {
+    // Reactotron.log("logout");
+    socket.emit("logout", token, (data) => {
+      if (data === "server error") {
+        dispatch({ type: actionTypes.LOG_OUT, logout: "false" });
+      }
+
+      if (data === "logout") {
+        localStorage.clear();
+        dispatch({ type: actionTypes.LOG_OUT, logout: "true" });
       }
     });
-
-    localStorage.clear();
-    dispatch({ type: actionTypes.LOG_OUT, logout: true });
   };
 }
